:recycle: Refactor : urls.py 내부 뷰 기능 함수 이동, model의 nickname 수정 (#13)
