+++ BE/users/models.py
@@ -31,7 +31,7 @@ class User(AbstractBaseUser, PermissionsMixin):
     objects = CustomUserManager()
 
     USERNAME_FIELD = 'email'
-    # REQUIRED_FIELDS = ['region', 'nickname']
+    REQUIRED_FIELDS = ['nickname']
 
     def __str__(self):
         return self.email
