+++ BE/users/urls.py
@@ -1,9 +1,4 @@
 from django.urls import path
-from django.contrib.auth.views import LoginView, LogoutView
-from rest_framework_simplejwt.tokens import RefreshToken
-from rest_framework.response import Response
-from rest_framework.decorators import api_view, permission_classes
-from rest_framework.permissions import IsAuthenticated
 
 from rest_framework_simplejwt.views import (
     TokenObtainPairView,
@@ -15,22 +10,8 @@ from .views import (
     UserDetailView,
     UserProfileEditView,
     UserDeactivateView,
+    user_logout,
 )
-
-@api_view(['POST'])
-@permission_classes([IsAuthenticated])
-def user_logout(request):
-    try:
-        # 현재 사용자의 refresh 토큰을 무효화
-        refresh_token = request.data.get('refresh', None)
-        if refresh_token:
-            RefreshToken(refresh_token).blacklist()
-
-        response_data = {'detail': 'Successfully logged out.'}
-        return Response(response_data, status=200)
-    except Exception as e:
-        response_data = {'detail': 'Failed to log out.'}
-        return Response(response_data, status=400)
 
 urlpatterns = [
     path('signup/', UserCreateView.as_view(), name='signup'), # 회원가입
