+++ BE/users/views.py
@@ -1,38 +1,57 @@
 from django.contrib.auth.mixins import LoginRequiredMixin
-from django.contrib.auth.views import LogoutView
 from django.shortcuts import render, redirect
 from django.urls import reverse_lazy
-from django.views.generic.edit import UpdateView, DeleteView
 from django.contrib.auth import get_user_model
 from rest_framework_simplejwt.views import TokenObtainPairView
-from rest_framework import permissions, generics
-
+from rest_framework import permissions, status
+from rest_framework.generics import CreateAPIView, RetrieveAPIView, UpdateAPIView, DestroyAPIView
+from rest_framework.response import Response
+from rest_framework.decorators import api_view, permission_classes
+from rest_framework.permissions import IsAuthenticated
+from rest_framework_simplejwt.tokens import RefreshToken
+from .models import User
 from .serializers import UserSerializer
 
-User = get_user_model()
 
-class UserCreateView(generics.CreateAPIView):
+class UserCreateView(CreateAPIView):
     queryset = User.objects.all()
     serializer_class = UserSerializer
     permission_classes = [permissions.AllowAny]
 
-class UserDetailView(generics.RetrieveAPIView):
+class UserDetailView(RetrieveAPIView):
     queryset = User.objects.all()
     serializer_class = UserSerializer
 
-class UserProfileEditView(LoginRequiredMixin, UpdateView):
-    model = User
-    # template_name = 'user_profile_edit.html'
-    fields = ['region', 'nickname', 'profile_image', 'profile_content']
-    success_url = reverse_lazy('user_detail')
+class UserProfileEditView(UpdateAPIView):
+    queryset = User.objects.all()
+    serializer_class = UserSerializer
+    permission_classes = [permissions.IsAuthenticated]
 
-    def get_object(self, queryset=None):
-        return self.request.user
+class UserDeactivateView(DestroyAPIView):
+    queryset = User.objects.all()
+    serializer_class = UserSerializer
+    permission_classes = [permissions.IsAuthenticated]
 
-class UserDeactivateView(LoginRequiredMixin, DeleteView):
-    model = User
-    # template_name = 'user_confirm_deactivate.html'
-    success_url = reverse_lazy('user_deactivate')
+    def destroy(self, request, *args, **kwargs):
+        instance = self.get_object()
+        # 사용자를 비활성화 처리하거나 삭제 로직을 추가할 수 있습니다.
+        instance.is_active = False
+        instance.save()
+        return Response(status=status.HTTP_204_NO_CONTENT)
+    
 
-    def get_object(self, queryset=None):
-        return self.request.user
+# view.py로 옮기기
+@api_view(['POST'])
+@permission_classes([IsAuthenticated])
+def user_logout(request):
+    try:
+        # 현재 사용자의 refresh 토큰을 무효화
+        refresh_token = request.data.get('refresh', None)
+        if refresh_token:
+            RefreshToken(refresh_token).blacklist()
+
+        response_data = {'detail': 'Successfully logged out.'}
+        return Response(response_data, status=200)
+    except Exception as e:
+        response_data = {'detail': 'Failed to log out.'}
+        return Response(response_data, status=400)
